Introduce an Ox::SyntaxError class (#233)
* Introduce an Ox::SyntaxError class

Ox currently raises a SyntaxError when certain invalid XML is encountered (for
instance, if a the document contains a prolog that is not the first element.)
Typically, SyntaxError is raised by the Ruby interpreter when code with invalid
syntax is encountered.  From that standpoint, it is an unexpected error for Ox
to be raising.  Additionally, SyntaxError does not inherit from StandardError,
meaning that it must be explicitly rescued (or Exception must be rescued) and a
bare rescue statement will not capture the error.

This change replaces Ox's use of SyntaxError with a new Ox::SyntaxError class
that does inherit from StandardError and will be more natural for developers to
work with.

* Add tests for dumping invalid characters and parsing unsupported Ox versions