Make each return an Enumerator
This allows for an easy use of Enumerable methods
like `#reduce` or `#map` by chaining them after
`#each` call.

Enumerable can be included to remove the extra call
but that would add methods to a class that intends
to use methods for navigation. Making `<reduce>` child
inaccessible with `#reduce` call, for example.