Add getter and setter for the indent value of `Ox::Builder` (#296)
* Add Ox::Builder#indent

This getter return the indent option of the builder.

* Add Ox::Builder#indent=

With this setter the indentation level can be changes dynamically. It
allows to build XML where some of the nodes are formatted inline within
the parent node.
This is useful for nodes that consider all spacing and line-breaking as
part of the value of the node's text.